Reconcile code and docstring for consistency_score and consistency (#195)
The documentation for the functions consistency_score and consistency
match the paper, which has a factor of 1 over k (k=n_neighbors in the
AIF360 documentation) to the left of the outer sum, rather than to the
left of the inner sum. The code in both functions does not match the
documentation since it takes the mean of y for the k nearest neighbours
(rather than the sum) and doesn't divide the outer sum by n_neighbors.

To see the error in the previous documentation, consider the extreme
case where y_hat = 1 for all examples. Here the consistency_score should
be 1 as implemented in the code.

Signed-off-by: Leena Murgai <code@leenamurgai.co.uk>